Created BookingService and BookingStatus (#16)

+++ src/main/java/com/pensionat/room/repository/RoomRepository.java
@@ -4,5 +4,5 @@ import com.pensionat.room.model.RoomEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomRepository
-        extends JpaRepository<RoomEntity, Integer> {
+        extends JpaRepository<RoomEntity, Long> {
 }
